# src/openapi.yaml
openapi: "3.0.3"
info:
  title: ScoutOff API
  description: >
    Backend API for ScoutOff — a decentralised football scouting platform
    built on the Stellar blockchain (Soroban smart contracts).
  version: "1.0.0"
  contact:
    name: ScoutOff Engineering
servers:
  - url: /api
    description: Stable API (version-less alias)
  - url: /api/v1
    description: Versioned API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Long-lived scout API key (alternative to a Bearer JWT). Restricted keys
        may only perform operations covered by their granted scopes (#1019).
        Legacy keys without a scope list are unrestricted.
  parameters:
    playerId:
      name: playerId
      in: path
      required: true
      schema:
        type: string
      description: Player's unique identifier (cuid2)
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string

    # ── Player ────────────────────────────────────────────────────────────────
    PlayerSummary:
      type: object
      required: [player_id, wallet]
      properties:
        player_id:
          type: string
        wallet:
          type: string
        position:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        metadataUri:
          type: string
          nullable: true
        progress_level:
          type: integer
          minimum: 0
          maximum: 3
        created_at:
          type: integer
          description: Unix timestamp (seconds)
        progress_tier_name:
          type: string
          enum: [Unverified, "Verified Identity", "Performance Milestones", "Elite Tier"]
        tierName:
          type: string
        tierDescription:
          type: string
        is_active:
          type: integer
          enum: [0, 1]
    PlayerDetail:
      allOf:
        - $ref: "#/components/schemas/PlayerSummary"
        - type: object
          required: [offerCount]
          properties:
            offerCount:
              type: integer
              minimum: 0
              description: Number of trial offers submitted for this player
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Insufficient role or wallet mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /auth/token:
    post:
      tags: [auth]
      summary: Exchange signed challenge for JWT
      operationId: postToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transaction, wallet, role]
              properties:
                transaction:
                  type: string
                  description: Signed SEP-10 XDR transaction
                wallet:
                  type: string
                role:
                  type: string
                  enum: [player, scout, validator, admin]
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Players ───────────────────────────────────────────────────────────────
...
  /players/register:
    post:
      tags: [players]
      summary: Register a new player profile
      operationId: registerPlayer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet, position, region]
              properties:
                wallet:
                  type: string
                  minLength: 56
                  maxLength: 56
                position:
                  type: string
                region:
                  type: string
                metadataUri:
                  type: string
                  description: Pre-pinned IPFS CID (mutually exclusive with metadata)
                metadata:
                  type: object
                  description: >
                    Arbitrary metadata object. The backend will pin it to IPFS
                    and derive the metadataUri.
      responses:
        "201":
          description: Player registered
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          playerId:
                            type: string
                          metadataUri:
                            type: string
                          gatewayUrl:
                            type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /players/{playerId}:
    get:
      tags: [players]
      summary: Get a player by ID
      operationId: getPlayer
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/playerId"
      responses:
        "200":
          description: Player detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PlayerDetail"
        "304":
          description: Not modified (ETag match)
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [players]
      summary: Update player profile metadata (owner only)
      operationId: updatePlayer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/playerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [metadataUri]
                  properties:
                    metadataUri:
                      type: string
                - type: object
                  required: [metadata]
                  properties:
                    metadata:
                      type: object
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          transactionId:
                            type: string
                          metadataUri:
                            type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /validators/milestone:
    post:
      tags: [validators]
      summary: Submit milestone evidence for a player
      operationId: submitMilestoneEvidence
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playerId, milestoneType, evidenceUri]
              properties:
                playerId:
                  type: string
                milestoneType:
                  type: string
                evidenceUri:
                  type: string
      responses:
        "201":
          description: Milestone evidence submitted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...